backend/requirements.txt
fastapi==0.115.12
uvicorn[standard]==0.34.0
sqlalchemy==2.0.39
asyncpg==0.30.0
alembic==1.15.1
pydantic-settings==2.8.1
structlog==25.2.0
email-validator==2.2.0
aiosqlite==0.21.0
python-multipart==0.0.20
pypdf==5.4.0
PyJWT==2.12.1
